HpetTimerDxe: HACK: use legacy IRQ to I/O APIC
On a system with an ICH10 southbridge, the standard routing to IRQ 20
does not seem to work. Possible reason:
"Reprogramming the HPET after switching to IOAPIC makes it work again",
see arch/x86/kernel/hpet.c:438 in the Linux 6.6.13 sources.

This hack assumes that Timer 0 of an ICH10 is used. The "legacy
replacement" routing of the interrupt delivers it to the I/O APIC on
IRQ 2. It is active high on this line, see ICH datasheet table 5-16,
note 1.

Linux uses this, too.